move authorized routes to own file
