+++ backend-express-app/api/index.js
@@ -6,6 +6,7 @@ const env = require("dotenv");
 const cors = require("cors");
 const { v4 } = require("uuid");
 const jwt = require("jsonwebtoken");
+// const authorizedRouter = require("../routes/authorized");
 
 //git pull origin a-exp -r 		pull updated branch, to rebase onto
 
@@ -17,17 +18,17 @@ var corsOptions = {
 	credentials: true,
 };
 
-const supabase = createClient(
-	process.env.SUPABASE_URL,
-	process.env.SUPABASE_ANON_KEY,
-	{
-		auth: {
-			autoRefreshToken: true,
-			persistSession: false,
-			detectSessionInUrl: true,
-		},
-	}
-);
+// const supabase = createClient(
+// 	process.env.SUPABASE_URL,
+// 	process.env.SUPABASE_ANON_KEY,
+// 	{
+// 		auth: {
+// 			autoRefreshToken: true,
+// 			persistSession: false,
+// 			detectSessionInUrl: true,
+// 		},
+// 	}
+// );
 
 const app = express();
 // app.use(cors({ origin: /test\-workout\-app\-vercel\.vercel\.app.*/ }));
@@ -69,7 +70,9 @@ app.use((req, res, next) => {
 // app.use(authMiddleware);
 
 //Routes
-app.use("/api/authorized", authorizedRouter);
+app.use("/api", require("../routes/index"));
+
+// app.use("/api/authorized", authorizedRouter);
 app.use("/api/public", publicRouter);
 
 // ROUTE Middleware = specific routes, like authorization
@@ -95,57 +98,57 @@ publicRouter.get("/api/item/:slug", (req, res) => {
 });
 
 // AUTHORIZED ROUTE
-//authorized route, will always take tokens and create/get a session
+// //authorized route, will always take tokens and create/get a session
 
-const setAuthorizedSessionMiddleware = async (req, res, next) => {
-	/* use req.cookies['cookie-name'] for COOKIES, but not vercel
-		const accessToken = req.cookies["my_access_token"];
-		const refreshToken = req.cookies["my_refresh_token"];
-		const user_id = req.cookies["my_user_is"]; */
+// const setAuthorizedSessionMiddleware = async (req, res, next) => {
+// 	/* use req.cookies['cookie-name'] for COOKIES, but not vercel
+// 		const accessToken = req.cookies["my_access_token"];
+// 		const refreshToken = req.cookies["my_refresh_token"];
+// 		const user_id = req.cookies["my_user_is"]; */
 
-	const refreshToken = req.headers["refresh-token"]; // do not just use req.cookies, turn into bearer tokens
-	const accessToken = req.headers["access-token"];
+// 	const refreshToken = req.headers["refresh-token"]; // do not just use req.cookies, turn into bearer tokens
+// 	const accessToken = req.headers["access-token"];
 
-	if (refreshToken && accessToken) {
-		const { data, error } = await supabase.auth.setSession({
-			refresh_token: refreshToken,
-			access_token: accessToken,
-		});
-		if (error) {
-			console.error("error fetching session from supabase: ", error);
-		}
-		if (data) {
-			console.log("session Set");
-		}
-	} else {
-		// make sure you handle this case!
-		console.error("User is not authenticated.");
-		const error = new Error(
-			"User is not authenticated! Access and/or refresh token needed"
-		);
-		console.log(error);
-		return res.status(401).send(error.message);
-	}
-	next();
-};
+// 	if (refreshToken && accessToken) {
+// 		const { data, error } = await supabase.auth.setSession({
+// 			refresh_token: refreshToken,
+// 			access_token: accessToken,
+// 		});
+// 		if (error) {
+// 			console.error("error fetching session from supabase: ", error);
+// 		}
+// 		if (data) {
+// 			console.log("session Set");
+// 		}
+// 	} else {
+// 		// make sure you handle this case!
+// 		console.error("User is not authenticated.");
+// 		const error = new Error(
+// 			"User is not authenticated! Access and/or refresh token needed"
+// 		);
+// 		console.log(error);
+// 		return res.status(401).send(error.message);
+// 	}
+// 	next();
+// };
 
-authorizedRouter.use(setAuthorizedSessionMiddleware);
+// authorizedRouter.use(setAuthorizedSessionMiddleware);
 
-// 	/api/authorized/workouts = All user's workouts
-authorizedRouter.get("/workouts", async (req, res) => {
-	console.log("Get All Workouts API");
-	// const userId = req.headers["user-id"];
+// // 	/api/authorized/workouts = All user's workouts
+// authorizedRouter.get("/workouts", async (req, res) => {
+// 	console.log("Get All Workouts API");
+// 	// const userId = req.headers["user-id"];
 
-	const { data, error } = await supabase.from("workouts").select("*");
-	if (data) {
-		console.log(data);
-		// if there is data, send it back = 200 status
-		res.send(data);
-	} else {
-		// if there is no data, send a 404 status with the err
-		res.status(404).send(error);
-	}
-});
+// 	const { data, error } = await supabase.from("workouts").select("*");
+// 	if (data) {
+// 		console.log(data);
+// 		// if there is data, send it back = 200 status
+// 		res.send(data);
+// 	} else {
+// 		// if there is no data, send a 404 status with the err
+// 		res.status(404).send(error);
+// 	}
+// });
 
 authorizedRouter.get("/workouts/:workoutUrl", async (req, res) => {
 	const workoutUrl = req.params.workoutUrl;
