+++ backend-express-app/routes/index.js
@@ -1,0 +1,5 @@
+// routes/index.js
+var router = require("express").Router();
+router.use("/authorized", require("./authorized")); // api/products/
+
+module.exports = router;
